ByteImage/requirement.txt
torchaudio
scipy
matplotlib
pillow
librosa
numpy<2.0
requests
bs4